requirements.txt
numpy
opencv-contrib-python
pafy
youtube_dl
imutils
pyqt5
Pillow
tqdm
tensorflow